+++ Packages/Core/Editor/Drawers/Constants/StringConstantDrawer.cs
@@ -7,6 +7,6 @@ namespace UnityAtoms.Editor
     /// Constant property drawer of type `string`. Inherits from `AtomDrawer&lt;StringConstant&gt;`. Only availble in `UNITY_2019_1_OR_NEWER`.
     /// </summary>
     [CustomPropertyDrawer(typeof(StringConstant))]
-    public class StringConstantDrawer : AtomDrawer<StringConstant> { }
+    public class StringConstantDrawer : VariableDrawer<StringConstant> { }
 }
 #endif