+++ Packages/Core/Editor/Drawers/Variables/BoolVariableDrawer.cs
@@ -7,6 +7,6 @@ namespace UnityAtoms.Editor
     /// Variable property drawer of type `bool`. Inherits from `AtomDrawer&lt;BoolVariable&gt;`. Only availble in `UNITY_2019_1_OR_NEWER`.
     /// </summary>
     [CustomPropertyDrawer(typeof(BoolVariable))]
-    public class BoolVariableDrawer : AtomDrawer<BoolVariable> { }
+    public class BoolVariableDrawer : VariableDrawer<BoolVariable> { }
 }
 #endif