+++ Packages/Core/Editor/Drawers/Variables/FloatVariableDrawer.cs
@@ -7,6 +7,6 @@ namespace UnityAtoms.Editor
     /// Variable property drawer of type `float`. Inherits from `AtomDrawer&lt;FloatVariable&gt;`. Only availble in `UNITY_2019_1_OR_NEWER`.
     /// </summary>
     [CustomPropertyDrawer(typeof(FloatVariable))]
-    public class FloatVariableDrawer : AtomDrawer<FloatVariable> { }
+    public class FloatVariableDrawer : VariableDrawer<FloatVariable> { }
 }
 #endif