+++ Packages/Core/Editor/Drawers/Variables/Vector3VariableDrawer.cs
@@ -7,6 +7,6 @@ namespace UnityAtoms.Editor
     /// Variable property drawer of type `Vector3`. Inherits from `AtomDrawer&lt;Vector3Variable&gt;`. Only availble in `UNITY_2019_1_OR_NEWER`.
     /// </summary>
     [CustomPropertyDrawer(typeof(Vector3Variable))]
-    public class Vector3VariableDrawer : AtomDrawer<Vector3Variable> { }
+    public class Vector3VariableDrawer : VariableDrawer<Vector3Variable> { }
 }
 #endif