+++ Packages/Core/Editor/Editors/Variables/IntVariableEditor.cs
@@ -1,0 +1,13 @@
+// generated file
+using UnityEditor;
+
+namespace UnityAtoms.Editor
+{
+    /// <summary>
+    /// Variable Inspector of type `int`. Inherits from `AtomVariableEditor`
+    /// </summary>
+    [CustomEditor(typeof(IntVariable))]
+    public sealed class IntVariableEditor : AtomVariableEditor
+    {
+    }
+}