+++ Packages/Core/Editor/Editors/Variables/Vector2VariableEditor.cs
@@ -1,0 +1,14 @@
+// generated file
+using UnityEditor;
+using UnityEngine;
+
+namespace UnityAtoms.Editor
+{
+    /// <summary>
+    /// Variable Inspector of type `Vector2`. Inherits from `AtomVariableEditor`
+    /// </summary>
+    [CustomEditor(typeof(Vector2Variable))]
+    public sealed class Vector2VariableEditor : AtomVariableEditor
+    {
+    }
+}