+++ Packages/Core/Editor/Editors/Variables/Vector3VariableEditor.cs
@@ -1,0 +1,14 @@
+// generated file
+using UnityEditor;
+using UnityEngine;
+
+namespace UnityAtoms.Editor
+{
+    /// <summary>
+    /// Variable Inspector of type `Vector3`. Inherits from `AtomVariableEditor`
+    /// </summary>
+    [CustomEditor(typeof(Vector3Variable))]
+    public sealed class Vector3VariableEditor : AtomVariableEditor
+    {
+    }
+}