+++ Packages/Core/Runtime/Listeners/FloatFloatListener.cs
@@ -6,7 +6,7 @@ namespace UnityAtoms
     /// Listener x 2 of type `float`. Inherits from `AtomListener&lt;float, float, FloatFloatAction, FloatFloatEvent, FloatFloatUnityEvent&gt;`.
     /// </summary>
     [EditorIcon("atom-icon-orange")]
-    [AddComponentMenu("Unity Atoms/Listeners/Float Float Listener")]
+    [AddComponentMenu("Unity Atoms/Listeners/Float x 2 Listener")]
     public sealed class FloatFloatListener : AtomListener<
         float,
         float,