+++ Packages/Core/Runtime/Listeners/IntIntListener.cs
@@ -6,7 +6,7 @@ namespace UnityAtoms
     /// Listener x 2 of type `int`. Inherits from `AtomListener&lt;int, int, IntIntAction, IntIntEvent, IntIntUnityEvent&gt;`.
     /// </summary>
     [EditorIcon("atom-icon-orange")]
-    [AddComponentMenu("Unity Atoms/Listeners/Int Int Listener")]
+    [AddComponentMenu("Unity Atoms/Listeners/Int x 2 Listener")]
     public sealed class IntIntListener : AtomListener<
         int,
         int,