+++ Packages/Mobile/Editor/Editors/Variables/TouchUserInputVariableEditor.cs
@@ -1,0 +1,15 @@
+// generated file
+using UnityEditor;
+using UnityAtoms.Mobile;
+using UnityAtoms.Editor;
+
+namespace UnityAtoms.Mobile.Editor
+{
+    /// <summary>
+    /// Variable Inspector of type `TouchUserInput`. Inherits from `AtomVariableEditor`
+    /// </summary>
+    [CustomEditor(typeof(TouchUserInputVariable))]
+    public sealed class TouchUserInputVariableEditor : AtomVariableEditor
+    {
+    }
+}