+++ CHANGELOG.md
@@ -4,17 +4,27 @@
 🏠 = Internal
 🏃‍♀= Performance
 💅 = Polish
-🚀 = New feature
+🚀 = New features
 
 # [Unreleased]
+
+## 🐛 Bug fixes
+
+-   [#91](https://github.com/AdamRamberg/unity-atoms/issues/91) Name OnTriggerHook class properly. ([@AdamRamberg](https://github.com/AdamRamberg))
 
 ## 📝 Documentation
 
 -   [#73](https://github.com/AdamRamberg/unity-atoms/issues/73) Add discord link to docs. ([@AdamRamberg](https://github.com/AdamRamberg))
+-   [#84](https://github.com/AdamRamberg/unity-atoms/pull/84) Changed AddComponentMenu names for Listeners/Hooks. ([@IceTrooper](https://github.com/IceTrooper))
 
-## 🚀 New feature
+## 🚀 New features
 
 -   [#78](https://github.com/AdamRamberg/unity-atoms/pull/78) Setter to AtomReference's Value was added. ([@Saso222](https://github.com/Saso222), [@soraphis](https://github.com/soraphis))
+
+## 💅 Polish
+
+-   [#70](https://github.com/AdamRamberg/unity-atoms/pull/70) Better user guidance for working with AtomVariables. ([@soraphis](https://github.com/soraphis), [@AdamRamberg](https://github.com/AdamRamberg))
+-   [#89](https://github.com/AdamRamberg/unity-atoms/pull/89) Variable and Constant Drawers show a preview value. ([@soraphis](https://github.com/soraphis), [@AdamRamberg](https://github.com/AdamRamberg))
 
 # 2.0.0 (October 24, 2019)